+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - added: Added Sepolia Ethereum testnet
 - added: Added support for WebSocket connections to Blockbook servers via `BlockbookWsAdapter`
+- remove: Remove deprecated Ethereum testnets: Goerli, Kovan, Rinkeby, and Ropsten.
 
 ## 4.13.0 (2024-07-16)